Fixes #3605 and #3600
