+++ field/FieldGet.F90
@@ -7,19 +7,19 @@ module mapl3g_FieldGet
    use mapl_ErrorHandling
    use mapl3g_UngriddedDims
    use esmf
-   implicit none (type, external)
+   implicit none (type,external)
    private
 
-   public :: MAPL_FieldGet
-   public :: MAPL_FieldSet
+   public :: FieldGet
+   public :: FieldSet
 
-   interface MAPL_FieldGet
+   interface FieldGet
       procedure field_get
-   end interface MAPL_FieldGet
+   end interface FieldGet
 
-   interface MAPL_FieldSet
+   interface FieldSet
       procedure field_set
-   end interface MAPL_FieldSet
+   end interface FieldSet
 
 contains
 
@@ -27,7 +27,8 @@ contains
         short_name, typekind, &
         num_levels, vert_staggerloc, num_vgrid_levels, &
         ungridded_dims, &
-        units, standard_name, long_name,  &
+        units, standard_name, long_name, &
+        is_connected, &
         rc)
 
       type(ESMF_Field), intent(in) :: field
@@ -41,6 +42,7 @@ contains
       character(len=:), optional, allocatable, intent(out) :: units
       character(len=:), optional, allocatable, intent(out) :: standard_name
       character(len=:), optional, allocatable, intent(out) :: long_name
+      logical, optional, intent(out) :: is_connected
 
       integer, optional, intent(out) :: rc
 
@@ -63,7 +65,9 @@ contains
            vert_staggerloc=vert_staggerloc, &
            num_vgrid_levels=num_vgrid_levels, &
            ungridded_dims=ungridded_dims, &
-           units=units, standard_name=standard_name, long_name=long_name, _RC)
+           units=units, standard_name=standard_name, long_name=long_name, &
+           is_connected=is_connected, &
+           _RC)
 
       _RETURN(_SUCCESS)
    end subroutine field_get
@@ -72,6 +76,7 @@ contains
    subroutine field_set(field, num_levels, vert_staggerloc, &
         ungridded_dims, &
         units, &
+        is_connected, &
         rc)
 
       type(ESMF_Field), intent(inout) :: field
@@ -79,6 +84,7 @@ contains
       type(VerticalStaggerLoc), optional, intent(in) :: vert_staggerloc
       type(UngriddedDims), optional, intent(in) :: ungridded_dims
       character(len=*), optional, intent(in) :: units
+      logical, optional, intent(in) :: is_connected
 
       integer, optional, intent(out) :: rc
 
@@ -91,7 +97,10 @@ contains
            num_levels=num_levels, &
            vert_staggerloc=vert_staggerloc, &
            ungridded_dims=ungridded_dims, &
-           units=units, _RC)
+           units=units, &
+           is_connected=is_connected, &
+           _RC)
+      
 
       _RETURN(_SUCCESS)
    end subroutine field_set
