+++ state/CMakeLists.txt
@@ -2,7 +2,9 @@ esma_set_this (OVERRIDE MAPL.state)
 
 set(srcs
   API.F90
-  StateGetPointerToData.F90
+  StateGet.F90
+  StateSet.F90
+  StateGetPointer.F90
 )
 
 list (APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}")
@@ -13,7 +15,7 @@ endif ()
 
 esma_add_library(${this}
   SRCS ${srcs}
-  DEPENDENCIES MAPL.shared ESMF::ESMF
+  DEPENDENCIES MAPL.field MAPL.shared MAPL.esmf_utils ESMF::ESMF
   TYPE SHARED
 )
