+++ state/undef.macro
@@ -1,0 +1,8 @@
+#undef IDENTITY_
+#undef DIMENSIONS_
+#undef SUB_
+#undef SUB__
+#undef SUB___
+#undef EKIND_
+#undef EKIND__
+#undef EKIND___
